@@ -1,10 +1,10 @@
 import React from 'react'
 import './App.css'
 
-import { HashRouter as Router, Route, Switch } from 'react-router-dom'
+import { HashRouter as Router, Route, Routes } from 'react-router-dom'
 
-import { createMuiTheme, ThemeProvider, responsiveFontSizes } from '@material-ui/core/styles'
-import CssBaseline from '@material-ui/core/CssBaseline'
+import { createTheme, ThemeProvider, responsiveFontSizes } from '@mui/material/styles'
+import CssBaseline from '@mui/material/CssBaseline'
 
 import AuthProvider, { AuthIsSignedIn, AuthIsNotSignedIn } from './contexts/authContext'
 
@@ -17,39 +17,39 @@ import ChangePassword from './routes/auth/changePassword'
 import Landing from './routes/landing'
 import Home from './routes/home'
 
-let lightTheme = createMuiTheme({
+let lightTheme = createTheme({
   palette: {
-    type: 'light',
+    mode: 'light',
   },
 })
 lightTheme = responsiveFontSizes(lightTheme)
 
-// let darkTheme = createMuiTheme({
+// let darkTheme = createTheme({
 //   palette: {
-//     type: 'dark',
+//     mode: 'dark',
 //   },
 // })
 // darkTheme = responsiveFontSizes(darkTheme)
 
 const SignInRoute: React.FunctionComponent = () => (
   <Router>
-    <Switch>
-      <Route path="/signin" component={SignIn} />
-      <Route path="/signup" component={SignUp} />
-      <Route path="/verify" component={VerifyCode} />
-      <Route path="/requestcode" component={RequestCode} />
-      <Route path="/forgotpassword" component={ForgotPassword} />
-      <Route path="/" component={Landing} />
-    </Switch>
+    <Routes>
+      <Route path="/signin" element={<SignIn />} />
+      <Route path="/signup" element={<SignUp />} />
+      <Route path="/verify" element={<VerifyCode />} />
+      <Route path="/requestcode" element={<RequestCode />} />
+      <Route path="/forgotpassword" element={<ForgotPassword />} />
+      <Route path="/" element={<Landing />} />
+    </Routes>
   </Router>
 )
 
 const MainRoute: React.FunctionComponent = () => (
   <Router>
-    <Switch>
-      <Route path="/changepassword" component={ChangePassword} />
-      <Route path="/" component={Home} />
-    </Switch>
+    <Routes>
+      <Route path="/changepassword" element={<ChangePassword />} />
+      <Route path="/" element={<Home />} />
+    </Routes>
   </Router>
 )
 
