@@ -1,6 +1,6 @@
 import React from 'react'
 
-import TextField from '@material-ui/core/TextField'
+import {TextField} from '@mui/material'
 
 export const Email: React.FunctionComponent<{ emailIsValid: boolean; setEmail: (_: string) => void }> = ({
   emailIsValid,
