@@ -1,28 +1,25 @@
-import React, { useState, useContext } from 'react'
+import { useState, useContext } from 'react'
 
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
-import { makeStyles } from '@material-ui/core/styles'
-import Box from '@material-ui/core/Box'
-import Grid from '@material-ui/core/Grid'
-import Button from '@material-ui/core/Button'
-import Typography from '@material-ui/core/Typography'
-import Paper from '@material-ui/core/Paper'
+import { styled } from '@mui/material/styles'
+import Box from '@mui/material/Box'
+import Grid from '@mui/material/Grid'
+import Button from '@mui/material/Button'
+import Typography from '@mui/material/Typography'
+import Paper from '@mui/material/Paper'
 
 import { useValidCode, useValidPassword, useValidUsername } from '../../hooks/useAuthHooks'
 import { Code, Password, Username } from '../../components/authComponents'
 
 import { AuthContext } from '../../contexts/authContext'
 
-const useStyles = makeStyles({
-  root: {
+const FullHeightDiv = styled(Grid)({
     height: '100vh',
-  },
-})
+});
+
 
 export default function ForgotPassword() {
-  const classes = useStyles()
-
   const { code, setCode, codeIsValid } = useValidCode('')
   const { password, setPassword, passwordIsValid } = useValidPassword('')
   const { username, setUsername, usernameIsValid } = useValidUsername('')
@@ -45,7 +42,7 @@ export default function ForgotPassword() {
     !passwordConfirmIsValid ||
     passwordConfirm.length === 0
 
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const authContext = useContext(AuthContext)
 
@@ -84,9 +81,9 @@ export default function ForgotPassword() {
 
       {/* Buttons */}
       <Box mt={2}>
-        <Grid container direction="row" justify="center">
+        <Grid container direction="row" justifyContent="center">
           <Box m={1}>
-            <Button onClick={() => history.goBack()} color="secondary" variant="contained">
+            <Button onClick={() => navigate(-1)} color="secondary" variant="contained">
               Cancel
             </Button>
           </Box>
@@ -105,7 +102,7 @@ export default function ForgotPassword() {
       <Typography variant="h5">{`Password Reset`}</Typography>
 
       <Box m={4}>
-        <Button onClick={() => history.push('/signin')} color="primary" variant="contained">
+        <Button onClick={() => navigate('/signin')} color="primary" variant="contained">
           Sign In
         </Button>
       </Box>
@@ -113,13 +110,13 @@ export default function ForgotPassword() {
   )
 
   return (
-    <Grid className={classes.root} container direction="row" justify="center" alignItems="center">
-      <Grid xs={11} sm={6} lg={4} container direction="row" justify="center" alignItems="center" item>
+    <FullHeightDiv container direction="row" justifyContent="center" alignItems="center">
+      <Grid xs={11} sm={6} lg={4} container direction="row" justifyContent="center" alignItems="center" item>
         <Paper style={{ width: '100%', padding: 16 }}>
-          <Grid container direction="column" justify="center" alignItems="center">
+          <Grid container direction="column" justifyContent="center" alignItems="center">
             {/* Title */}
             <Box m={3}>
-              <Grid container direction="row" justify="center" alignItems="center">
+              <Grid container direction="row" justifyContent="center" alignItems="center">
                 <Typography variant="h3">Forgot Password</Typography>
               </Grid>
             </Box>
@@ -128,6 +125,6 @@ export default function ForgotPassword() {
           </Grid>
         </Paper>
       </Grid>
-    </Grid>
+    </FullHeightDiv>
   )
 }
