@@ -1,33 +1,28 @@
-import React, { useState, useContext } from 'react'
+import { useState, useContext } from 'react'
 
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
-import { makeStyles } from '@material-ui/core/styles'
-import Box from '@material-ui/core/Box'
-import Grid from '@material-ui/core/Grid'
-import Button from '@material-ui/core/Button'
-import Typography from '@material-ui/core/Typography'
-import Paper from '@material-ui/core/Paper'
+import { styled } from '@mui/material/styles'
+import Box from '@mui/material/Box'
+import Grid from '@mui/material/Grid'
+import Button from '@mui/material/Button'
+import Typography from '@mui/material/Typography'
+import Paper from '@mui/material/Paper'
 
 import { useValidUsername } from '../../hooks/useAuthHooks'
 import { Username } from '../../components/authComponents'
 
 import { AuthContext } from '../../contexts/authContext'
 
-const useStyles = makeStyles({
-  root: {
+const FullHeightRoot = styled(Grid)({
     height: '100vh',
-  },
-  hover: {
-    '&:hover': { cursor: 'pointer' },
-  },
-  text: {
-    textAlign: 'center',
-  },
-})
+});
+
+const TypographyWithCenterText = styled(Typography)({
+  textAlign: 'center'
+});
 
 export default function RequestCode() {
-  const classes = useStyles()
 
   const { username, setUsername, usernameIsValid } = useValidUsername('')
   const [error, setError] = useState('')
@@ -35,7 +30,7 @@ export default function RequestCode() {
 
   const isValid = !usernameIsValid || username.length === 0
 
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const authContext = useContext(AuthContext)
 
@@ -51,10 +46,10 @@ export default function RequestCode() {
   const emailSent = (
     <>
       <Box mt={1}>
-        <Typography className={classes.text} variant="h5">{`Reset Code Sent to ${username}`}</Typography>
+        <TypographyWithCenterText variant="h5">{`Reset Code Sent to ${username}`}</TypographyWithCenterText>
       </Box>
       <Box mt={4}>
-        <Button onClick={() => history.push('forgotpassword')} color="primary" variant="contained">
+        <Button onClick={() => navigate('/forgotpassword')} color="primary" variant="contained">
           Reset Password
         </Button>
       </Box>
@@ -73,9 +68,9 @@ export default function RequestCode() {
       </Box>
 
       <Box mt={2}>
-        <Grid container direction="row" justify="center">
+        <Grid container direction="row" justifyContent="center">
           <Box m={1}>
-            <Button color="secondary" variant="contained" onClick={() => history.goBack()}>
+            <Button color="secondary" variant="contained" onClick={() => navigate(-1)}>
               Cancel
             </Button>
           </Box>
@@ -90,10 +85,10 @@ export default function RequestCode() {
   )
 
   return (
-    <Grid className={classes.root} container direction="row" justify="center" alignItems="center">
-      <Grid xs={11} sm={6} lg={4} container direction="row" justify="center" alignItems="center" item>
+    <FullHeightRoot container direction="row" justifyContent="center" alignItems="center">
+      <Grid xs={11} sm={6} lg={4} container direction="row" justifyContent="center" alignItems="center" item>
         <Paper style={{ width: '100%', padding: 32 }}>
-          <Grid container direction="column" justify="center" alignItems="center">
+          <Grid container direction="column" justifyContent="center" alignItems="center">
             <Box m={2}>
               <Typography variant="h3">Send Reset Code</Typography>
             </Box>
@@ -102,6 +97,6 @@ export default function RequestCode() {
           </Grid>
         </Paper>
       </Grid>
-    </Grid>
+    </FullHeightRoot>
   )
 }
