@@ -1,27 +1,24 @@
 import React, { useState, useContext } from 'react'
 
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
-import { makeStyles } from '@material-ui/core/styles'
-import Box from '@material-ui/core/Box'
-import Grid from '@material-ui/core/Grid'
-import Button from '@material-ui/core/Button'
-import Typography from '@material-ui/core/Typography'
-import Paper from '@material-ui/core/Paper'
+import { styled } from '@mui/material/styles'
+import Box from '@mui/material/Box'
+import Grid from '@mui/material/Grid'
+import Button from '@mui/material/Button'
+import Typography from '@mui/material/Typography'
+import Paper from '@mui/material/Paper'
 
 import { useValidEmail, useValidPassword, useValidUsername } from '../../hooks/useAuthHooks'
 import { Email, Password, Username } from '../../components/authComponents'
 
 import { AuthContext } from '../../contexts/authContext'
 
-const useStyles = makeStyles({
-  root: {
+const FullHeightRoot = styled(Grid)({
     height: '100vh',
-  },
-})
+});
 
 const SignUp: React.FunctionComponent<{}> = () => {
-  const classes = useStyles()
 
   const { email, setEmail, emailIsValid } = useValidEmail('')
   const { password, setPassword, passwordIsValid } = useValidPassword('')
@@ -45,7 +42,7 @@ const SignUp: React.FunctionComponent<{}> = () => {
     !passwordConfirmIsValid ||
     passwordConfirm.length === 0
 
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const authContext = useContext(AuthContext)
 
@@ -82,9 +79,9 @@ const SignUp: React.FunctionComponent<{}> = () => {
 
       {/* Buttons */}
       <Box mt={2}>
-        <Grid container direction="row" justify="center">
+        <Grid container direction="row" justifyContent="center">
           <Box m={1}>
-            <Button onClick={() => history.goBack()} color="secondary" variant="contained">
+            <Button onClick={() => navigate(-1)} color="secondary" variant="contained">
               Cancel
             </Button>
           </Box>
@@ -104,7 +101,7 @@ const SignUp: React.FunctionComponent<{}> = () => {
       <Typography variant="h6">{`Verfiy Code sent to ${email}`}</Typography>
 
       <Box m={4}>
-        <Button onClick={() => history.push('/verify')} color="primary" variant="contained">
+        <Button onClick={() => navigate('/verify')} color="primary" variant="contained">
           Send Code
         </Button>
       </Box>
@@ -112,13 +109,13 @@ const SignUp: React.FunctionComponent<{}> = () => {
   )
 
   return (
-    <Grid className={classes.root} container direction="row" justify="center" alignItems="center">
-      <Grid xs={11} sm={6} lg={4} container direction="row" justify="center" alignItems="center" item>
+    <FullHeightRoot container direction="row" justifyContent="center" alignItems="center">
+      <Grid xs={11} sm={6} lg={4} container direction="row" justifyContent="center" alignItems="center" item>
         <Paper style={{ width: '100%', padding: 16 }}>
-          <Grid container direction="column" justify="center" alignItems="center">
+          <Grid container direction="column" justifyContent="center" alignItems="center">
             {/* Title */}
             <Box m={3}>
-              <Grid container direction="row" justify="center" alignItems="center">
+              <Grid container direction="row" justifyContent="center" alignItems="center">
                 <Typography variant="h3">Sign Up</Typography>
               </Grid>
             </Box>
@@ -127,7 +124,7 @@ const SignUp: React.FunctionComponent<{}> = () => {
           </Grid>
         </Paper>
       </Grid>
-    </Grid>
+    </FullHeightRoot>
   )
 }
 
