@@ -1,30 +1,29 @@
 import React, { useState, useContext } from 'react'
 
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
-import { makeStyles } from '@material-ui/core/styles'
-import Box from '@material-ui/core/Box'
-import Grid from '@material-ui/core/Grid'
-import Button from '@material-ui/core/Button'
-import Typography from '@material-ui/core/Typography'
-import Paper from '@material-ui/core/Paper'
+import { styled } from '@mui/material/styles'
+import Box from '@mui/material/Box'
+import Grid from '@mui/material/Grid'
+import Button from '@mui/material/Button'
+import Typography from '@mui/material/Typography'
+import Paper from '@mui/material/Paper'
 
 import { useValidCode, useValidUsername } from '../../hooks/useAuthHooks'
 import { Code, Username } from '../../components/authComponents'
 
 import { AuthContext } from '../../contexts/authContext'
 
-const useStyles = makeStyles({
-  root: {
+const FullHeightRoot = styled(Grid)({
     height: '100vh',
-  },
-  hover: {
+});
+
+const TypographyWithHover = styled(Typography)({
     '&:hover': { cursor: 'pointer' },
-  },
-})
+  });
+
 
 const VerifyCode: React.FunctionComponent<{}> = () => {
-  const classes = useStyles()
 
   const { username, setUsername, usernameIsValid } = useValidUsername('')
   const { code, setCode, codeIsValid } = useValidCode('')
@@ -32,28 +31,28 @@ const VerifyCode: React.FunctionComponent<{}> = () => {
 
   const isValid = !usernameIsValid || username.length === 0 || !codeIsValid || code.length === 0
 
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const authContext = useContext(AuthContext)
 
   const sendClicked = async () => {
     try {
       await authContext.verifyCode(username, code)
-      history.push('signin')
+      navigate('signin')
     } catch (err) {
       setError('Invalid Code')
     }
   }
 
   const passwordResetClicked = async () => {
-    history.push('/resetpassword')
+    navigate('/resetpassword')
   }
 
   return (
-    <Grid className={classes.root} container direction="row" justify="center" alignItems="center">
-      <Grid xs={11} sm={6} lg={4} container direction="row" justify="center" alignItems="center" item>
+    <FullHeightRoot container direction="row" justifyContent="center" alignItems="center">
+      <Grid xs={11} sm={6} lg={4} container direction="row" justifyContent="center" alignItems="center" item>
         <Paper style={{ width: '100%', padding: 32 }}>
-          <Grid container direction="column" justify="center" alignItems="center">
+          <Grid container direction="column" justifyContent="center" alignItems="center">
             {/* Title */}
             <Box m={2}>
               <Typography variant="h3">Send Code</Typography>
@@ -66,11 +65,11 @@ const VerifyCode: React.FunctionComponent<{}> = () => {
             </Box>
             <Box width="80%" m={1}>
               <Code codeIsValid={codeIsValid} setCode={setCode} />
-              <Grid container direction="row" justify="flex-start" alignItems="center">
+              <Grid container direction="row" justifyContent="flex-start" alignItems="center">
                 <Box onClick={passwordResetClicked} mt={2}>
-                  <Typography className={classes.hover} variant="body2">
+                  <TypographyWithHover variant="body2">
                     Resend Code
-                  </Typography>
+                  </TypographyWithHover>
                   <Box mt={2}>
                     <Typography color="error" variant="body2">
                       {error}
@@ -82,9 +81,9 @@ const VerifyCode: React.FunctionComponent<{}> = () => {
 
             {/* Buttons */}
             <Box mt={2}>
-              <Grid container direction="row" justify="center">
+              <Grid container direction="row" justifyContent="center">
                 <Box m={1}>
-                  <Button color="secondary" variant="contained" onClick={() => history.goBack()}>
+                  <Button color="secondary" variant="contained" onClick={() => navigate(-1)}>
                     Cancel
                   </Button>
                 </Box>
@@ -98,7 +97,7 @@ const VerifyCode: React.FunctionComponent<{}> = () => {
           </Grid>
         </Paper>
       </Grid>
-    </Grid>
+    </FullHeightRoot>
   )
 }
 
